+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fields `userCreated`, `dateCreated`, `userModified`, `dateModified` are saved correctly in all write methods
 
 ## [6.3.0] - 2022-05-06
 ### Added